// src/scrapers/orientation-ch/types.ts
export interface OrientationChProfession {
  name: string;
  url: string;
  orientationId: string;
  domainesProfessionnels: string;
  niveauxDeFormation: string;
  swissdoc: string;
  miseAJour: string;
  description: string;
  formation: string;
  perspectivesProfessionnelles: string;
  autresInformations: string;
  adressesUtiles: string;
}